refactor:フィルタ実行部分を削除
move:Sobelフォルダに移動
